fix: address review feedback on error wrapping, collision messages, predicates, and metric types
- Wrap root cause error in validatePolicyRefExists for both
  RestrictedRoleDefinition and RestrictedBindDefinition webhooks
  instead of dropping it with errors.New
- Include conflicting RestrictedRoleDefinition name in cross-type
  targetName collision error messages in RoleDefinition webhook
- Add GenerationChangedPredicate to RBACPolicy controller watches
  for RestrictedBindDefinition and RestrictedRoleDefinition to
  avoid unnecessary reconcile churn on status-only updates
- Change SetPolicyViolationsActive violations parameter from
  float64 to int for type safety; convert at the Prometheus
  .Set() boundary

+++ internal/controller/authorization/restricted_helpers.go
@@ -91,7 +91,7 @@ func handlePolicyViolations(
 	conditions.MarkFalse(obj, authorizationv1alpha1.PolicyCompliantCondition, generation,
 		authorizationv1alpha1.PolicyCompliantReasonViolationsDetected, authorizationv1alpha1.PolicyCompliantMessageViolationsDetected, strings.Join(msgStrings, "; "))
 
-	metrics.SetPolicyViolationsActive(cfg.ControllerLabel, runtimeObj.GetName(), float64(len(violations)))
+	metrics.SetPolicyViolationsActive(cfg.ControllerLabel, runtimeObj.GetName(), len(violations))
 
 	recorder.Eventf(runtimeObj, nil, corev1.EventTypeWarning,
 		authorizationv1alpha1.EventReasonPolicyViolation, authorizationv1alpha1.EventActionReconcile,
